src/Modelo/Bean/ExoPagoV.java

package Modelo.Bean;

public class ExoPagoV {
    private String CodigoE;
    private String fecha;
    private String Placa;
    private String fechaDeuda;
    private String FechaSolicitud;
    private float monto;
    private String estado;

    public String getCodigoE() {
        return CodigoE;
    }

    public void setCodigoE(String CodigoE) {
        this.CodigoE = CodigoE;
    }

    public String getFecha() {
        return fecha;
    }

    public void setFecha(String fecha) {
        this.fecha = fecha;
    }

    public String getPlaca() {
        return Placa;
    }

    public void setPlaca(String Placa) {
        this.Placa = Placa;
    }

    public String getFechaDeuda() {
        return fechaDeuda;
    }

    public void setFechaDeuda(String fechaDeuda) {
        this.fechaDeuda = fechaDeuda;
    }

    public String getFechaSolicitud() {
        return FechaSolicitud;
    }

    public void setFechaSolicitud(String FechaSolicitud) {
        this.FechaSolicitud = FechaSolicitud;
    }

    

    public String getEstado() {
        return estado;
    }

    public void setEstado(String estado) {
        this.estado = estado;
    }

    public float getMonto() {
        return monto;
    }

    public void setMonto(float monto) {
        this.monto = monto;
    }
}
